new evaluation logic

+++ src/Logic/Calcu_Logic.vala
@@ -18,21 +18,45 @@ public class Evaluation:GLib.Object
     public Evaluation(config c=config(){use_degrees=true})
     {
         this.update(c);
+        snd_evaluation = new Evaluation.secondary();
+        //test values TODO remove
+        //this.fun_extern.data[0] = new UserFuncData.with_data("p*x", {"x"});
+        //this.fun_extern.data[1] = new UserFuncData.with_data("sqrt(xx+yy)", {"x", "y"});
+        //this.fun_extern.data[2] = new UserFuncData.with_data("x+y+xy", {"x", "y"});
+    }
+
+    private Evaluation.secondary() {
+
+    }
+
+    public Evaluation.small(config c = config(){use_degrees = true}) {
+        this.update(c);
+    }
+
+    public Evaluation.with_data(GenericArray<Part?> parts, GenericArray<Sequence?> seq) {
+        var _section = new GenericArray<Part?>();
+        var _sequence = new GenericArray<Sequence?>();
+
+        parts.foreach( (x) => _section.add(x) );
+        seq.foreach( (x) => _sequence.add(x) );
+
+        this.section = _section;
+        this.sequence = _sequence;
     }
 
     public void update(config c) {
         fun_intern=get_intern_functions(c.use_degrees);
+        fun_extern = get_extern_functions(c.custom_functions);
         operator=get_operator();
         variable=get_custom_variable(c.custom_variable);
         con=c;
     }
 
-    public config con;
-
+    public config con{get; set;}
+    public Evaluation? snd_evaluation = null;
 	private PreparePart[] parts={};
 	public string input{get; set; default="";}
 	public double? result{get; private set; default=null;}
-	//public string? s_result{get; private set; default=null;}
 	private GenericArray<Part?> section=new GenericArray<Part?>();
 	private GenericArray<Sequence?>sequence=new GenericArray<Sequence?>();
 
@@ -41,13 +65,26 @@ public class Evaluation:GLib.Object
 
 	public Operation operator{get; set;}
     public Func fun_intern{get; set; }
-	public string[] control{get; set; default={"(",")",","," "};}
+    public UserFunc fun_extern {get; set;}
+	public string[] control{get; set; default={"(", ")", ",", " "};}
 	public Replaceable variable{get; set;}
 
     public void clear(){
         parts={};
         section.remove_range(0,section.length);
         sequence.remove_range(0,sequence.length);
+    }
+
+    public PreparePart[] get_parts() {
+        return this.parts;
+    }
+
+    public GenericArray<Part?> get_section() {
+        return this.section;
+    }
+
+    public GenericArray<Sequence?> get_sequence() {
+        return this.sequence;
     }
 
 	public void split() throws CALC_ERROR
@@ -60,6 +97,7 @@ public class Evaluation:GLib.Object
 		var cnt=PreparePart(){type=Type.CONTROL};
 		var vrb=PreparePart(){type=Type.VARIABLE};
 		var fui=PreparePart(){type=Type.EXPRESSION};
+		var fue=PreparePart(){type=Type.FUNCTION};
 		PreparePart ap={};
 
 		int len_num;
@@ -68,6 +106,7 @@ public class Evaluation:GLib.Object
 		int ind_cnt=-1;
 		int ind_vrb=-1;
 		int ind_fui=-1;
+		int ind_fue=-1;
 
 		while(input.length>0) {
 
@@ -76,36 +115,40 @@ public class Evaluation:GLib.Object
 			cnt.value=next_match(input,control, out ind_cnt);
 			vrb.value=next_match(input,variable.key, out ind_vrb);
 			fui.value=next_match(input,fun_intern.key, out ind_fui);
+			fue.value=next_match(input,fun_extern.key, out ind_fue);
 
 			opr.length=opr.value.length;
 			cnt.length=cnt.value.length;
 			vrb.length=vrb.value.length;
 			fui.length=fui.value.length;
+			fue.length=fue.value.length;
 			num.length=len_num;
 
 			opr.index=ind_op;
 			cnt.index=ind_cnt;
 			vrb.index=ind_vrb;
             fui.index=ind_fui;
+            fue.index=ind_fue;
 
-			ap=get_longest(opr,num,cnt,vrb,fui);
+			ap=get_longest(opr,num,cnt,vrb,fui,fue);
 
 			if(ap.length>0)
 			    {
                     if(check_mul&&(!(ap.type==Type.OPERATOR||ap.type==Type.NUMBER||(ap.type==Type.CONTROL&&!(ap.value=="(")))))
                         parts+=PreparePart(){value="*", type=Type.OPERATOR, length=1, index=2};
 
+                   if (ap.value == "-" && (parts.length == 0 || (parts[parts.length - 1].type == Type.CONTROL && parts[parts.length - 1].value != ")")))
+                        parts += PreparePart(){value="0", type=Type.NUMBER};
+
 			        parts+=ap;
 			        input=input[ap.length:input.length];
 
-			        can_negative=(ap.type==Type.OPERATOR||(ap.type==Type.CONTROL&&ap.value!=")"))?true:false;
-			        check_mul=(ap.type==Type.NUMBER||ap.type==Type.VARIABLE)?true:false;
+			        can_negative=(ap.type==Type.OPERATOR||(ap.type==Type.CONTROL&&ap.value!=")"));
+			        check_mul=(ap.type==Type.NUMBER||ap.type==Type.VARIABLE||ap.value==")");
 			    }
 			else {
 			    throw new CALC_ERROR.INVALID_SYMBOL (@"the symbol `$(input[0:1])` is not known");
 			}
-
-
 
 		}
 	}
@@ -121,14 +164,16 @@ public class Evaluation:GLib.Object
 			{
 				case Type.VARIABLE: {
 					section.add(Part(){
-						value=variable.value[part.index]
+						value=variable.value[part.index],
+						has_value = true
 					});
 
 					break;
 				}
 				case Type.NUMBER: {
 					section.add(Part() {
-						value=double.parse(part.value)
+						value=double.parse(part.value),
+						has_value = true
 					});
 
 					break;
@@ -157,11 +202,26 @@ public class Evaluation:GLib.Object
 
 				    break;
 				}
+				case Type.FUNCTION: {
+				    section.add(Part(){
+				        eval = fun(){eval = fun_extern.eval, arg_right = fun_extern.arg_right[part.index]},
+				        data  = (fun_extern.data[part.index]).with_evaluation(snd_evaluation)
+				    });
+				    //TODO pass config
+				    sequence.add(Sequence(){
+				        priority = 4 + bracket_value,
+				        arguments = fun_extern.arg_right[part.index],
+				        index = section.length - 1
+				    });
+				    break;
+				}
 				case Type.CONTROL: {
-					if(part.value==")")
-					bracket_value-=bracket;
-					else if(part.value=="(")
-					bracket_value+=bracket;
+					if(part.value==")") {
+					    bracket_value-=bracket;
+					}
+					else if(part.value=="(") {
+					    bracket_value+=bracket;
+					}
 					break;
 				}
 				default: {
@@ -184,53 +244,15 @@ public class Evaluation:GLib.Object
 
 	public void eval() throws CALC_ERROR
 	{
-//
-//
-//
-	    int score=0;
-	    int promised=0;
-	    int changes=1;
-	    bool control=false;
-	    Part sec;
 
-	    for (int i=0; i<section.length; i++) {
-	    sec=section.get(i);
-	        changes=1;
-	        control=false;
-            if(sec.value==null)
-            {
-                promised+=sec.eval.arg_right;
-                changes=1-sec.eval.arg_left;
-                control=true;
-            }
-
-                if(promised>0&&score!=0)
-                    promised-=changes;
-                else
-                    score+=changes;
-                if(control&&score!=1) {
-                if(score<1)
-                    throw new CALC_ERROR.MISSING_ARGUMENT("Missing Argument, a left argument is required");
-                else
-                    throw new CALC_ERROR.REMAINING_ARGUMENT("Remaining Argument");
-                    }
-	    }
-
-        if(promised!=0)
-            throw new CALC_ERROR.MISSING_ARGUMENT("Missing Argument");
-        if(score!=1) {
-                if(score<1)
-                    throw new CALC_ERROR.MISSING_ARGUMENT("Missing Argument");
-                else
-                    throw new CALC_ERROR.REMAINING_ARGUMENT("Remaining Argument");
-            }
-//
-//
-//
 		//sortierung nach priority
+		//int64 msec = GLib.get_real_time();
 		sequence.sort(sorting);
+		//stdout.printf(@"__time:$( (GLib.get_real_time()-msec)/1000 )\n");
 		//index berechnung
+		//msec = GLib.get_real_time();
 		sequence=eval_seq(sequence);
+		//stdout.printf(@"__time:$( (GLib.get_real_time()-msec)/1000 )\n");
 
 
 		for(int i=0; i<sequence.length; i++)
@@ -243,21 +265,30 @@ public class Evaluation:GLib.Object
 			//get arg_left
 			for(int l=0; l<part.eval.arg_left; l++)
 			{
-				arg+=section.get(ind-part.eval.arg_left).value;
-				section.remove_index(ind-part.eval.arg_left);
+			    if ( (ind - part.eval.arg_left) >= 0 && section.get(ind - part.eval.arg_left).has_value) {
+				    arg+=section.get(ind-part.eval.arg_left).value;
+				    section.remove_index(ind-part.eval.arg_left);
+				}
+				else throw new CALC_ERROR.MISSING_ARGUMENT(@"Missing Argument, '$(parts[ind].value)' requires a left argument");
 			}
 
 			//get arg_right
 			for(int l=0; l<part.eval.arg_right; l++)
 			{
-				arg+=section.get(ind+1-part.eval.arg_left).value;
-				section.remove_index(ind+1-part.eval.arg_left);
+			    if ( (ind + 1 - part.eval.arg_left) < section.length && section.get(ind + 1 - part.eval.arg_left).has_value) {
+				    arg+=section.get(ind+1-part.eval.arg_left).value;
+				    section.remove_index(ind+1-part.eval.arg_left);
+				}
+				else throw new CALC_ERROR.MISSING_ARGUMENT(@"Missing Argument, '$(parts[ind].value)' requires $(part.eval.arg_right) right $( (part.eval.arg_right > 1) ? "arguments" : "argument"  )");
 			}
 
 			section.set(ind-part.eval.arg_left,Part(){
-				value=part.eval.eval(arg)
+				value=part.eval.eval(arg, part.data),
+				has_value = true
 			});
 		}
+		if (section.length > 1)
+		    throw new CALC_ERROR.REMAINING_ARGUMENT(@"$(section.length - 1) $( (section.length > 2) ? "arguments are" : "argument is" ) remaining");
 		result=section.get(0).value??null;
 		//s_result=result.to_string();
 		if(con.round_decimal) {
@@ -266,7 +297,7 @@ public class Evaluation:GLib.Object
 	    }
 	}
 
-    public double eval_auto(string in,config? c=null) throws CALC_ERROR {
+    public double eval_auto(string in, config? c=null) throws CALC_ERROR {
         this.input=in;
         if(c!=null)
             this.update(c);
@@ -309,11 +340,32 @@ public class Evaluation:GLib.Object
 		return ret;
 	}
 
-	// zu genie <
+	// to genie <
 	CompareFunc<Sequence?> sorting = (a, b) => {
 		return (int) (a.priority < b.priority) - (int) (a.priority > b.priority);
 	};
 	// >
+    public static  Eval fun_extern_eval = (value, data) =>{
+                    var func_data = data as UserFuncData;
+
+                    for (int i = 0; i < func_data.parts.length; i++)
+                        func_data.evaluation.section.add(func_data.parts[i]);
+                    for (int i = 0; i < func_data.sequence.length; i++)
+                        func_data.evaluation.sequence.add(func_data.sequence[i]);
+
+                    // causes "g_object_ref: assertion 'G_IS_OBJECT (object)' failed"
+                    //func_data.parts.foreach((part) => func_data.evaluation.section.add(part));
+                    //func_data.sequence.foreach((seq) => func_data.evaluation.sequence.add(seq));
+
+                    for (int i = 0; i < func_data.part_index.length; i++) {
+                        func_data.evaluation.section[func_data.part_index[i]].value = value[func_data.argument_index[i]];
+                    }
+
+                    func_data.evaluation.eval();
+                    func_data.evaluation.clear();
+                    return func_data.evaluation.result ?? 0 / 0;
+            };
+
 }
 
 }
